make updates on Describe Publication page non-reloading
The add/remove buttons on the Describe Publication page cause
a full-page reload, submitting the entire Publication Metadata
form to the server and receiving an updated page in return.

This commit adds per-item, non-reloading updates for adding
and removing metadata for the page, using an AJAX request
and a partial-page update instead of a full page reload
for add/remove actions. The two Save buttons continue to
cause a page reload, since they take the user away from
the Describe Publication page.

This update affects only the form with @id=
and does not affect the payment form.

The update in this commit is to the file lib/js/dryad-pages.js
in the Mirage theme, and it adds a number of JS functions
to bind to the submit events and isolate the page fragment
from the request response to use in updating the DOM. The
update makes heavy use of jQuery-provided functionality
for node selection and for the AJAX request.