on response to describe-publication form, update entire form
Instead of replacing only part of the form once an AJAX response
has been received from the server, replace the entire form element.
A whole-form update maintains a non-reload refresh of the page
without adding any special case code around the contriubtor and
corresponding-author linkage, which may be updated in the back
end.